refactor: refactor components sub components gf-233
