+++ apps/frontend/src/libs/components/sidebar/libs/components/components.ts
@@ -1,0 +1,1 @@
+export { SidebarItem } from "./sidebar-item/sidebar-item.js";
