+++ apps/frontend/src/libs/components/table/libs/components/select-row-cell/select-row-cell.tsx
